requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.7.1
google-generativeai==0.3.2
python-multipart==0.0.6
requests==2.31.0
torch==2.1
transformers==4.34.0
ctransformers
python-dotenv
langchain
chromadb
pypdf2
sentence-transformers
tiktoken
tqdm
together
langchain-together
langchain_community
pymongo[srv]==3.12
langgraph==0.0.69
numpy==1.24.3
PyPDF2==3.0.1
tavily-python==0.5.0